Move the proc macro to a separated crate
