+++ soa-derive-internal/src/slice.rs
