+++ soa-derive-internal/src/structs.rs
